+++ Engine.Core.UnitTests/SetUpFixture.cs
@@ -3,6 +3,7 @@ using NUnit.Framework;
 using System;
 using System.Collections.Generic;
 using System.Text;
+using Engine.Core;
 
 namespace Engine.Core.Tests
 {