+++ Engine.Core/GameApplication.cs
@@ -2,10 +2,12 @@
 using Engine.Core.Managers.Interfaces;
 using Engine.Core.Models.Enums;
 using Engine.Core.Models.Interfaces;
+
 using Microsoft.Extensions.Logging;
 using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
 using Microsoft.Xna.Framework.Input;
+
 using System;
 using System.Collections.Concurrent;
 using System.Collections.Generic;
@@ -23,6 +25,9 @@ namespace Engine.Core
         protected readonly DiagnosticsController _diagnosticsController;
         protected readonly ILogger _logger;
         protected SpriteBatch _spriteBatch;
+
+        protected string GameTitle { get; set; } = "Game";
+        private bool GameRunning { get; set; } = true;
 
         private ConcurrentQueue<ConcurrentBag<IEntity>> DrawState { get; set; }
 
@@ -45,6 +50,9 @@ namespace Engine.Core
         protected virtual Task InitialiseAsync()
         {
             base.Initialize();
+            // Get engine advertisment screen on there
+            // Screen set up
+
             return Task.CompletedTask;
         }
 
@@ -52,6 +60,8 @@ namespace Engine.Core
         {
             base.LoadContent();
             _spriteBatch = new SpriteBatch(GraphicsDevice);
+            // Set up all the models?
+
             return Task.CompletedTask;
         }
 
@@ -63,57 +73,83 @@ namespace Engine.Core
 
         protected virtual async Task UpdateAsync(GameTime gameTime)
         {
-            await _eventManager.ProcessEvent();
             var entitiesToEnqueue = await _entityManager.UpdateEntities(gameTime);
-
             // Not equal to update because we only want to draw entities which can be drawn
             var bag = new ConcurrentBag<IEntity>(entitiesToEnqueue.Where(x => x.EntityLifeCycleAction() != EntityActions.UPDATE));
 
             DrawState.Enqueue(bag);
+
             base.Update(gameTime);
         }
 
         protected virtual async Task DrawAsync(GameTime gameTime)
         {
-            List<Task> renderTasks = new List<Task>();
             if (DrawState.TryDequeue(out ConcurrentBag<IEntity> entitiesToDraw))
-            {
-                _logger.LogDebug("Drawing {DrawCount} entities", entitiesToDraw.Count);
-                foreach (var entity in entitiesToDraw)
-                    renderTasks.ToList().Add(entity.Render(_spriteBatch, gameTime));
-
-                await Task.WhenAll(renderTasks);
-            }
+                await _entityManager.DrawEntities(gameTime, entitiesToDraw.ToList());
             else
                 _logger.LogError("Attempted to dequeue from DrawState but returned false");
 
+            await _entityManager.DrawEntities(gameTime);
+
             base.Draw(gameTime);
+        }
+
+        public void RunGame()
+        {
+            try
+            {
+                Run();
+            }
+            catch (Exception ex)
+            {
+                _logger.LogError(ex, "{MethodName} caught an exception. Ending application early.", nameof(RunGame));
+                throw;
+            }            
+        }
+
+        protected override void OnActivated(object sender, EventArgs args)
+        {
+            Window.Title = GameTitle;
+            base.OnActivated(sender, args);
+        }
+
+        protected override void OnDeactivated(object sender, EventArgs args)
+        {
+            Window.Title = $"{GameTitle} | Deactive";
+            // Open Menu pause
+            base.OnDeactivated(sender, args);
+        }
+
+        protected override void OnExiting(object sender, EventArgs args)
+        {
+            GameRunning = false;
+            base.OnExiting(sender, args);
         }
 
         #region Game Implementation
         protected override async void Initialize()
         {
-            await _diagnosticsController.DiagnoseAsyncMethod(InitialiseAsync(), nameof(InitialiseAsync));
+            await _diagnosticsController.DiagnoseTask(InitialiseAsync(), nameof(InitialiseAsync));
         }
 
         protected override async void LoadContent()
         {
-            await _diagnosticsController.DiagnoseAsyncMethod(LoadContentAsync(), nameof(LoadContentAsync));
+            await _diagnosticsController.DiagnoseTask(LoadContentAsync(), nameof(LoadContentAsync));
         }
 
         protected override async void UnloadContent()
         {
-            await _diagnosticsController.DiagnoseAsyncMethod(UnloadContentAsync(), nameof(UnloadContentAsync));
+            await _diagnosticsController.DiagnoseTask(UnloadContentAsync(), nameof(UnloadContentAsync));
         }
 
         protected override async void Update(GameTime gameTime)
         {
-            await _diagnosticsController.DiagnoseAsyncMethod(UpdateAsync(gameTime), nameof(UpdateAsync));
+            await _diagnosticsController.DiagnoseTask(UpdateAsync(gameTime), nameof(UpdateAsync));
         }
 
         protected override async void Draw(GameTime gameTime)
         {
-            await _diagnosticsController.DiagnoseAsyncMethod(DrawAsync(gameTime), nameof(DrawAsync));
+            await _diagnosticsController.DiagnoseTask(DrawAsync(gameTime), nameof(DrawAsync));
         }
         #endregion Game Implementation
     }