+++ Engine.Core/Managers/Interfaces/IEventManager.cs
@@ -9,8 +9,8 @@ namespace Engine.Core.Managers.Interfaces
 {
     public interface IEventManager
     {
+        void Begin();
         Task PublishEvent<T>(T @event) where T : MyEvent;
-        Task ProcessEvent();
         Task SubscribeToEvent(EventType eventType, IEventSubscriber subscriber);
         Task UnsubscribeFromEvent(EventType eventType, IEventSubscriber subscriber);
     }