+++ ExampleGame/Entities/Triangle.cs
@@ -14,15 +14,5 @@ namespace ExampleGame.Entities
         public Triangle()
         {
         }
-
-        public override Task Render(SpriteBatch spriteBatch, GameTime gameTime)
-        {
-            return base.Render(spriteBatch, gameTime);
-        }
-
-        public override Task Update(GameTime gameTime)
-        {
-            return base.Update(gameTime);
-        }
     }
 }