+++ ExampleGame/GameInitialiser.cs
@@ -1,10 +1,16 @@
-﻿namespace ExampleGame
+﻿using Microsoft.Extensions.DependencyInjection;
+using System;
+using System.Threading.Tasks;
+
+namespace ExampleGame
 {
     public static class GameInitialiser
     {
         public static void Main()
         {
-            Engine.Core.Engine.Run<Game1>();
+            var services = new ServiceCollection().AddMyOwnServices();
+
+            Engine.Core.Engine.Run<Game1>(services);
         }
     }
 }