+++ ExampleGame/ServiceExtensions.cs
@@ -15,7 +15,8 @@ namespace ExampleGame
         {
             // Transient for splitscreen, singleton for 1 entity for all players
             services
-                .AddSingleton<ICamera, Camera>();
+                .AddSingleton<ICamera, Camera>()
+                .AddSingleton<TestService>();
 
             return services;
         }