+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 ### Monogame Engine Changelog
 
-* 28/07/2019 - Duncan Baldwin
+* 24/08/2019 - dunkybaldy
+  * Set up event handling to run on a separate thread.
+  * Task.Factory.StartNew(action)
+  * Diagnose high level functions
+  * Allow game title easy change
+  * Simple start up from executable main method
+
+* 28/07/2019 - dunkybaldy
   * Set up base engine with entity management
   * Example test suite
   * Async Await functionality